requirements.txt
#Bootstrap_Flask
#Flask_CKEditor
#Flask_Login
#Flask-Gravatar
#Flask_WTF
#WTForms
#Werkzeug
flask
python-dotenv
psycopg2-binary
sqlalchemy
